separate events and data

+++ src/client/components/player.rs
@@ -2,11 +2,11 @@ use bevy::prelude::*;
 
 #[derive(Component)]
 pub struct Player {
-    pub id: u8,
+    pub id: usize,
 }
 
 impl Player {
-    pub fn new(id: u8) -> Self {
+    pub fn new(id: usize) -> Self {
         Self { id }
     }
 }
